Expose create_option() and use for proforma options; base MutuallyExclusiveOption classes on AusTrakkaCliOption
