+++ austrakka/utils/misc.py
@@ -51,24 +51,6 @@ class AusTrakkaCliTopLevel(click.Group):
             sys.exit(exc.exit_code)
 
 
-def _get_custom_help_record(orig_help, multiple):
-    if multiple:
-        tmp_list = list(orig_help)
-        split_str = tmp_list[len(tmp_list) - 1].rsplit("]", 1)
-        if len(split_str) > 1:
-            tmp_list[len(tmp_list) - 1] = split_str[0] + ";Accepts Multiple]"
-        else:
-            tmp_list[len(tmp_list) - 1] += " [Accepts Multiple]"
-        orig_help = tuple(tmp_list)
-    return orig_help
-
-
-class AusTrakkaCliOption(click.Option):
-    def get_help_record(self, ctx):
-        orig_help = super().get_help_record(ctx)
-        return _get_custom_help_record(orig_help, self.multiple)
-
-
 def logger_wraps(*, entry=True, exit_func=True):
 
     def wrapper(func):
